Refactor email services
EmailService has been split on several subservices, each of them having
a single responsability:

- EmailNotificationListener acts as a controller. Its responsability is
to handle and preparing any incoming request/notification for sending an
email. It  prepares any data needed for sending an email, delegating
into EmailService for doing the sending.

- EmailService / SimpleJavaMailService has the responsability of sending
an email using the input data. It is helped by TemplateService in order
to generate HTML body email if needed.

- TemplateService / ThymeleafTemplateService has the responsability of
instantiating any input templates and variables, generating the final
HTML body for an email.